application/Backend/SubscriptionBasedEcommerceApp/src/main/java/com/hsbc/model/Category.java
package com.hsbc.model;

public class Category {
    private int categoryId;
    private String categoryName;
    private String categoryDescription;

    public Category() {
    }


    public Category(String categoryName, int categoryId, String categoryDescription) {
        this.categoryName = categoryName;
        this.categoryId = categoryId;
        this.categoryDescription = categoryDescription;
    }

    public int getCategoryId() {
        return categoryId;
    }

    public void setCategoryId(int categoryId) {
        this.categoryId = categoryId;
    }

    public String getCategoryName() {
        return categoryName;
    }

    public void setCategoryName(String categoryName) {
        this.categoryName = categoryName;
    }

    public String getCategoryDescription() {
        return categoryDescription;
    }

    public void setCategoryDescription(String categoryDescription) {
        this.categoryDescription = categoryDescription;
    }

    @Override
    public String toString() {
        return "Category{" +
                "categoryId=" + categoryId +
                ", categoryName='" + categoryName + '\'' +
                ", categoryDescription='" + categoryDescription + '\'' +
                '}';
    }
}

